% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_baumarten}
\alias{tg_baumarten}
\title{Baumarten der TreeGrOSS-Kollektion}
\usage{
tg_baumarten()
}
\value{
Ein Dataframe mit allen in der TreeGrOSS-Kollektion verf체gbaren
Baumarten mit den Spalten \code{code}, \code{kurz}, \code{name}, \code{latein}.
}
\description{
Listet alle in der TreeGrOSS-Kollektion verf체gbaren Baumarten. Angegeben
sind Baumartencode, K체rzel, deutscher und lateinischer Name.
Jede dieser Angaben kann in allen \code{tg_}-Funktionen verwendet werden, um eine
Baumart auszuw채hlen. Die Suche nach der passenden Baumart erfolgt in
der oben genannten Reihenfolge.
}
\examples{
tg_baumarten()

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}},
\code{\link[=tg_kronenansatz]{tg_kronenansatz()}}, \code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}},
\code{\link[=tg_fun_info]{tg_fun_info()}}
}
