% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_bonitaet100}
\alias{tg_bonitaet100}
\title{Oberhöhenbonität (TreeGrOSS)}
\usage{
tg_bonitaet100(ba, alter, h100, info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{alter}{numerischer Vektor der Baumalter in Jahren. Eines oder mehrere.}

\item{h100}{Oberhöhe, Höhe des Grundflächenmittelstammes der
100 stärksten Bäume je Hektar der betrachteten Kohorte, z.B. Baumart oder
Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Oberhöhenbonitäten in Meter, wenn
\code{info=FALSE}. Ansonsten ein Listen-Objekt mit den Elementen \code{werte},
\code{formel}, \code{quelle} und \code{info}, welche Oberhöhenbonitäten, die verwendete
Formel, die Quellenangabe (Zitat) und ggf. Informationen über einen
Ersatz enthält.
}
\description{
Baumweise Bonität als Oberhöhebonität mit Bezugsalter 100 in Meter.
}
\details{
Die Funktion nutzt das Element \code{"SiteIndex"} aus der TreeGrOSS-XML-Datei
\code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016). Sie wurde für
Nordwestdeutschland parametrisiert.
}
\examples{
tg_bonitaet100(211, alter=150, h100=40)
tg_bonitaet100(111, alter=100, h100=27.6, info=TRUE)
tg_bonitaet100('Esche', alter=50:55, h100=26)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}},
\code{\link[=tg_kronenansatz]{tg_kronenansatz()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}}, \code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
