% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_fun_info}
\alias{tg_fun_info}
\title{Infos über Funktionen der TreeGrOSS-Kollektion}
\usage{
tg_fun_info(ba, fun)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{fun}{Name einer TreeGrOSS Funktion. Eine von \code{"tg_hoehe"},
\code{"tg_volumen"}, \code{"tg_kronenbreite"}, \code{"tg_kronenansatz"},
\code{"tg_bonitaet100"}, \code{"tg_hoehenzuwachs"}}
}
\value{
Ein Listen-Objekt mit den Elementen \code{formel}, \code{quelle} und \code{info},
welche die verwendete Formel, die Quellenangabe (Zitat) und ggf.
Informationen über einen Ersatz enthält.
}
\description{
Zeigt die Formel, einen Hinweis zur Quelle und Informationen über
baumartenweise Ersatz der Formel, denn nicht alle Baumarten haben eigene
Formeln. Manchmal wird die Formel einer anderen Baumart von TreeGrOSS als
Ersatz für die angefragte Baumart verwendet. Dies wird in der Regel im
\code{info}-Abschnittvermerkt.
}
\details{
In der TreeGrOSS XML-Datei werden Ersatzformeln auf zwei Arten definiert.
Erstens durch Angabe einer Referenz auf eine andere Art (z.B.
\verb{HandledLikeCode 211}). Alle \code{tg_}-Funktionen werten diese Referenzen aus
und teilen die Ersetzung im \code{info}-Abschnitt mit. Zweitens durch Kopieren und
Einfügen der Formel von einer Baumart zu einer anderen Baumart. In diesem
Fall kann die Ersetzung im \code{quelle}-Abschnitt sichtbar sein (z.B.
\verb{Eiche (NAGEL 1999)}) oder auch nicht.
}
\examples{
tg_fun_info(ba=211, fun="tg_volumen")
tg_fun_info(ba="Traubeneiche", fun="tg_hoehenzuwachs")

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}},
\code{\link[=tg_kronenansatz]{tg_kronenansatz()}}, \code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}},
\code{\link[=tg_baumarten]{tg_baumarten()}}
}
