% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_hoehe}
\alias{tg_hoehe}
\title{Baumhöhen aus Einheitshöhenkurve (TreeGrOSS)}
\usage{
tg_hoehe(ba, bhd, dg, hg, info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{bhd}{numerischer Vektor der Brusthöhendurchmesser in Zentimeter.
Einer oder mehrere.}

\item{dg}{Durchmesser des Grundflächenmittelstammes der betrachteten Kohorte,
z.B. Baumart oder Stichprobenpunkt, in Zentimeter. Nur ein Wert.}

\item{hg}{Höhe des Grundflächenmittelstammes der betrachteten Kohorte,
z.B. Baumart oder Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Baumhöhen in Meter, wenn \code{info=FALSE}.
Ansonsten ein Listen-Objekt mit den Elementen \code{werte}, \code{formel},
\code{quelle} und \code{info}, welche ein vektor der Baumhöhen, die verwendete
Formel, die Quellenangabe (Zitat) und ggf. Informationen über einen
Ersatz enthält.
}
\description{
Baumhöhen aus baumartenspezifischen Einheitshöhenkurven mit Parametern aus
TreeGrOSS.
}
\details{
Die Funktion nutzt das Element \code{"UniformHeightCurveXML"} aus der
TreeGrOSS-XML-Datei \code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016).
Sie wurde für Nordwestdeutschland parametrisiert.
}
\examples{
tg_hoehe(ba=211, bhd=32, hg=29.3, dg=29.4)
tg_hoehe(511, bhd=12:16, hg=19, dg=17.5)
tg_hoehe(ba="Stieleiche", bhd=60, hg=33.4, dg=63.2, info=TRUE)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}}, \code{\link[=tg_kronenansatz]{tg_kronenansatz()}},
\code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}}, \code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
