% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_hoehenzuwachs}
\alias{tg_hoehenzuwachs}
\title{Höhenzuwachs (TreeGrOSS)}
\usage{
tg_hoehenzuwachs(ba, alter, h, bon100, h100, info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{alter}{numerischer Vektor der Baumalter in Jahren. Eines oder mehrere.}

\item{h}{numerischer Vektor der Baumhöhen in Meter (vgl. \code{\link[=tg_hoehe]{tg_hoehe()}}).
Eine oder mehrere.}

\item{bon100}{numerischer Vektor mit Oberhöhenbonitäten in Meter. Oberhöhe
zum Bezugsalter 100 (vgl. \code{\link[=tg_bonitaet100]{tg_bonitaet100()}}). Eine oder mehrere.}

\item{h100}{Oberhöhe, Höhe des Grundflächenmittelstammes der
100 stärksten Bäume je Hektar der betrachteten Kohorte, z.B. Baumart oder
Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Höhenzuwächse für 5-Jahres Perioden in
Meter, wenn \code{info=FALSE}. Ansonsten ein Listen-Objekt mit den Elementen
\code{werte}, \code{formel}, \code{quelle} und \code{info}, welche ein Vektor der
Höhenzuwächse, die verwendete Formel, die Quellenangabe (Zitat) und ggf.
Informationen über einen Ersatz enthält.
}
\description{
Höhenzuwachs für 5-Jahresperioden in Meter.
}
\details{
Die Funktion nutzt das Element \code{"PotentialHeightIncrement"} aus der
TreeGrOSS-XML-Datei \code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016).
Sie wurde für Nordwestdeutschland parametrisiert.
}
\examples{
tg_hoehenzuwachs(511, alter=50, h=23, bon100=35.5, h100=23)
tg_hoehenzuwachs('Douglas', alter=30, h=20, bon100=46, h100=20.6)
tg_hoehenzuwachs(522, alter=120, h=37, bon100=35.4, h100=37.3, info=TRUE)
tg_hoehenzuwachs(211, alter=50:54, h=16:20, bon100=rep(33.5,5), h100=19.5)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}},
\code{\link[=tg_kronenansatz]{tg_kronenansatz()}}, \code{\link[=tg_bonitaet100]{tg_bonitaet100()}},
\code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
