% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_kronenansatz}
\alias{tg_kronenansatz}
\title{Kronenansatzhöhe (TreeGrOSS)}
\usage{
tg_kronenansatz(ba, bhd, h, h100, info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{bhd}{numerischer Vektor der Brusthöhendurchmesser in Zentimeter.
Einer oder mehrere.}

\item{h}{numerischer Vektor der Baumhöhen in Meter (vgl. \code{\link[=tg_hoehe]{tg_hoehe()}}).
Eine oder mehrere.}

\item{h100}{Oberhöhe, Höhe des Grundflächenmittelstammes der
100 stärksten Bäume je Hektar der betrachteten Kohorte, z.B. Baumart oder
Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Kronenansatzhöhen in Meter, wenn
\code{info=FALSE}. Ansonsten ein Listen-Objekt mit den Elementen \code{werte},
\code{formel}, \code{quelle} und \code{info}, welche ein Vektor der Kronenansatzhöhen,
die verwendete Formel, die Quellenangabe (Zitat) und ggf. Informationen
über einen Ersatz enthält.
}
\description{
Höhe des untersten Teils der lebenden Krone (z.B. unterster Grünast oder
unterster Quirl mit drei grünen Ästen).
}
\details{
Die Funktion nutzt das Element \code{"Crownbase"} aus der TreeGrOSS-XML-Datei
\code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016). Sie wurde für
Nordwestdeutschland parametrisiert.
}
\examples{
tg_kronenansatz(211, bhd=44, h=36, h100=38)
tg_kronenansatz(111, bhd=34, h=26, h100=28, info=TRUE)
tg_kronenansatz('Lärche', bhd=c(31,33,34), h=c(28,29,30), h100=32)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}},
\code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}}, \code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
