% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_kronenbreite}
\alias{tg_kronenbreite}
\title{Kronenbreite (TreeGrOSS)}
\usage{
tg_kronenbreite(ba, bhd, ..., info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{bhd}{numerischer Vektor der Brusthöhendurchmesser in Zentimeter.
Einer oder mehrere.}

\item{...}{um für die Erlenarten (420-423) die zusätzlichen Parameter
\code{h} und \code{ka}zu übergeben.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Kronenbreiten in Meter, wenn
\code{info=FALSE}. Ansonsten ein Listen-Objekt mit den Elementen \code{werte},
\code{formel}, \code{quelle} und \code{info}, welche ein Vektor der Kronenbreiten, die
verwendete Formel, die Quellenangabe (Zitat) und ggf. Informationen über
einen Ersatz enthält.
}
\description{
Durchmesser der Baumkrone in Abhängigkeit vom Brusthöhendurchmesser.
Die Erlenarten (Code 420-423) benötigen die zusätzlich Parameter Baumhöhe \code{h}
und Kronenansatz \code{ka} (vgl. \code{\link[=tg_kronenansatz]{tg_kronenansatz()}}). Es wird angenommen,
dass die  Kronen kreisförmig sind.
}
\details{
Die Funktion nutzt das Element \code{"Crownwidth"} aus der TreeGrOSS-XML-Datei
\code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016). Sie wurde für
Nordwestdeutschland parametrisiert.
}
\examples{
tg_kronenbreite(211, 20)
tg_kronenbreite('Buche', 20, info=TRUE)
tg_kronenbreite('Alnus incana', bhd=40, h=29, ka=10.2)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_volumen]{tg_volumen()}}, \code{\link[=tg_kronenansatz]{tg_kronenansatz()}},
\code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}}, \code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
