% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_parameter}
\alias{tg_parameter}
\title{Alle Parameter der TreeGrOSS-Kollektion}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{bhd}{numerischer Vektor der Brusthöhendurchmesser in Zentimeter.
Einer oder mehrere.}

\item{h}{numerischer Vektor der Baumhöhen in Meter (vgl. \code{\link[=tg_hoehe]{tg_hoehe()}}).
Eine oder mehrere.}

\item{alter}{numerischer Vektor der Baumalter in Jahren. Eines oder mehrere.}

\item{bon100}{numerischer Vektor mit Oberhöhenbonitäten in Meter. Oberhöhe
zum Bezugsalter 100 (vgl. \code{\link[=tg_bonitaet100]{tg_bonitaet100()}}). Eine oder mehrere.}

\item{dg}{Durchmesser des Grundflächenmittelstammes der betrachteten Kohorte,
z.B. Baumart oder Stichprobenpunkt, in Zentimeter. Nur ein Wert.}

\item{hg}{Höhe des Grundflächenmittelstammes der betrachteten Kohorte,
z.B. Baumart oder Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{h100}{Oberhöhe, Höhe des Grundflächenmittelstammes der
100 stärksten Bäume je Hektar der betrachteten Kohorte, z.B. Baumart oder
Stichprobenpunkt, in Meter. Nur ein Wert.}

\item{fun}{Name einer TreeGrOSS Funktion. Eine von \code{"tg_hoehe"},
\code{"tg_volumen"}, \code{"tg_kronenbreite"}, \code{"tg_kronenansatz"},
\code{"tg_bonitaet100"}, \code{"tg_hoehenzuwachs"}}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}

\item{TestParameter}{Sollte auf keiner der tg-Hilfeseiten sichtbar sein.}
}
\description{
Diese Hilfeseite existiert hauptsächlich damit alle \code{tg_}-Funktionen
einheitliche Parameterbeschreibungen erhalten. Technisch erfolgt dies in
roxygen durch Vererbung (\code{"@inheritParams"}).
}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\keyword{internal}
