% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treegross.R
\name{tg_volumen}
\alias{tg_volumen}
\title{Derbholzvolumen (TreeGrOSS)}
\usage{
tg_volumen(ba, bhd, h, info = FALSE)
}
\arguments{
\item{ba}{Name der Baumart (Kürzel, deutscher Name, lateinischer Name) oder
Baumartencode (nach niedersächsischer Verschlüsselung). Nur ein Wert.
Für vollständige Liste siehe \code{\link[=tg_baumarten]{tg_baumarten()}}.}

\item{bhd}{numerischer Vektor der Brusthöhendurchmesser in Zentimeter.
Einer oder mehrere.}

\item{h}{numerischer Vektor der Baumhöhen in Meter (vgl. \code{\link[=tg_hoehe]{tg_hoehe()}}).
Eine oder mehrere.}

\item{info}{boolescher Wert. Sollen begleitende Informationen ausgegeben
werden? Für weitere Details siehe \code{\link[=tg_fun_info]{tg_fun_info()}}.}
}
\value{
Ein numerischer Vektor der Derbholzvolumen in Kubikmeter, wenn
\code{info=FALSE}. Ansonsten ein Listen-Objekt mit den Elementen \code{werte},
\code{formel}, \code{quelle} und \code{info}, welche ein Vektor der Derbholzvolumen, die
verwendete Formel, die Quellenangabe (Zitat) und ggf. Informationen über
einen Ersatz enthält.
}
\description{
Volumen der oberirdischen Holzmasse ab 7cm Durchmesser in Kubikmeter.
}
\details{
Die Funktion nutzt das Element \code{"VolumeFunctionXML"} aus der
TreeGrOSS-XML-Datei \code{ForestSimulatorNWGermany*.xml} (Nagel, 2006, 2016).
Sie wurde für Nordwestdeutschland parametrisiert.
}
\examples{
tg_volumen(ba='Fagus syl', bhd=52.3, h=37.8)
tg_volumen(511, c(15.5, 23.0, 30.7, 37.6), c(16.6, 24.7, 29.7, 33.3))
tg_volumen('Alnus incana', bhd=29.0, h=24.7, info=TRUE)

}
\references{
Nagel, J.; Duda, H.; Hansen, J. (2006): Forest Simulator BWINPro7.
Forst und Holz 61(10): 427-429.

Nagel, J. (2021): TreeGrOSS. \url{https://www.nw-fva.de/index.php?id=477}.
}
\seealso{
\code{\link[=tg_hoehe]{tg_hoehe()}}, \code{\link[=tg_kronenbreite]{tg_kronenbreite()}}, \code{\link[=tg_kronenansatz]{tg_kronenansatz()}},
\code{\link[=tg_bonitaet100]{tg_bonitaet100()}}, \code{\link[=tg_hoehenzuwachs]{tg_hoehenzuwachs()}}, \code{\link[=tg_fun_info]{tg_fun_info()}}, \code{\link[=tg_baumarten]{tg_baumarten()}}
}
